fix: add /promotions/active endpoint and convert BigDecimal to Double

+++ src/main/java/com/cinect/dto/response/PromotionResponse.java
@@ -4,7 +4,6 @@ import com.cinect.entity.enums.DiscountType;
 import com.cinect.entity.enums.PromotionStatus;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -15,9 +14,9 @@ public class PromotionResponse {
     private String description;
     private String code;
     private DiscountType discountType;
-    private BigDecimal discountValue;
-    private BigDecimal minPurchase;
-    private BigDecimal maxDiscount;
+    private Double discountValue;
+    private Double minPurchase;
+    private Double maxDiscount;
     private Integer usageLimit;
     private Integer usageCount;
     private Instant startDate;
